requirements.txt
chainlit==1.0.200
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.5.3
langchain==0.1.6
langchain-openai==0.0.6
langchain-community==0.0.20
faiss-cpu==1.7.4
pypdf2==3.0.1
python-docx==1.1.0
python-multipart==0.0.6
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
pandas==2.2.0
numpy==1.26.3
flask==2.3.3
flask-cors==4.0.0
aiohttp==3.8.5
pyyaml==6.0.1
psycopg2-binary==2.9.9
redis==5.0.1
werkzeug==2.3.7
gunicorn==21.2.0